requirements.txt
google-generativeai
openai
anthropic
pillow
replicate
tensorboardX
matplotlib
tensorboard